Split out lowering modules
